+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/presencetracing/organizer/TraceLocationOrganizerSettings.kt
@@ -12,10 +12,4 @@ class TraceLocationOrganizerSettings @Inject constructor(val preferences: TraceL
             return preferences.qrInfoAcknowledged.value
         }
         set(value) = preferences.qrInfoAcknowledged.update { value }
-
-    var warnInfoAcknowledged: Boolean
-        get() {
-            return preferences.warnInfoAcknowledged.value
-        }
-        set(value) = preferences.warnInfoAcknowledged.update { value }
 }
