+++ Corona-Warn-App/src/main/res/layout/trace_location_organizer_warn_info_fragment.xml
@@ -4,9 +4,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:contentDescription="@string/trace_location_organiser_warn_info_header"
     android:background="@color/colorSurface"
-    tools:context=".ui.presencetracing.organizer.warn.info.TraceLocationWarnInfoFragment">
+    android:contentDescription="@string/trace_location_organiser_warn_info_header"
+    tools:context=".ui.presencetracing.organizer.warn.TraceLocationWarnInfoFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbar"
@@ -22,10 +22,10 @@
         android:id="@+id/scrollview"
         android:layout_width="match_parent"
         android:layout_height="0dp"
+        app:layout_constraintBottom_toTopOf="@id/next_button"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/toolbar"
-        app:layout_constraintBottom_toTopOf="@id/next_button">
+        app:layout_constraintTop_toBottomOf="@id/toolbar">
 
 
         <androidx.constraintlayout.widget.ConstraintLayout
@@ -38,9 +38,9 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:contentDescription="@string/trace_location_qr_info_content_description"
-                app:srcCompat="@drawable/ic_trace_location_warn_info_illustration"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
+                app:layout_constraintTop_toTopOf="parent"
+                app:srcCompat="@drawable/ic_trace_location_warn_info_illustration" />
 
             <TextView
                 android:id="@+id/trace_location_warn_info_headline"
@@ -74,9 +74,9 @@
                 android:layout_height="wrap_content"
                 android:layout_marginTop="24dp"
                 android:importantForAccessibility="no"
-                app:srcCompat="@drawable/ic_qr_tracing_static"
                 app:layout_constraintStart_toStartOf="@id/trace_location_warn_info_subtitle"
-                app:layout_constraintTop_toBottomOf="@id/trace_location_warn_info_subtitle" />
+                app:layout_constraintTop_toBottomOf="@id/trace_location_warn_info_subtitle"
+                app:srcCompat="@drawable/ic_qr_tracing_static" />
 
             <TextView
                 android:id="@+id/trace_location_warn_info_guests"
@@ -96,9 +96,9 @@
                 android:layout_height="wrap_content"
                 android:layout_marginTop="24dp"
                 android:importantForAccessibility="no"
-                app:srcCompat="@drawable/ic_bureaucracy_illustration"
                 app:layout_constraintStart_toStartOf="@id/trace_location_tracing_icon"
-                app:layout_constraintTop_toBottomOf="@id/trace_location_warn_info_guests" />
+                app:layout_constraintTop_toBottomOf="@id/trace_location_warn_info_guests"
+                app:srcCompat="@drawable/ic_bureaucracy_illustration" />
 
             <TextView
                 android:id="@+id/trace_location_qr_code_illustration_text"
@@ -109,10 +109,10 @@
                 android:layout_marginEnd="@dimen/spacing_normal"
                 android:layout_marginBottom="40dp"
                 android:text="@string/trace_location_organiser_warn_info_authority"
+                app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toEndOf="@id/trace_location_qr_code_icon"
-                app:layout_constraintTop_toTopOf="@id/trace_location_qr_code_icon"
-                app:layout_constraintBottom_toBottomOf="parent"/>
+                app:layout_constraintTop_toTopOf="@id/trace_location_qr_code_icon" />
 
         </androidx.constraintlayout.widget.ConstraintLayout>
     </ScrollView>
@@ -122,15 +122,15 @@
         style="@style/buttonPrimary"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/spacing_small"
         android:layout_marginStart="@dimen/spacing_normal"
+        android:layout_marginTop="@dimen/spacing_small"
         android:layout_marginEnd="@dimen/spacing_normal"
         android:layout_marginBottom="@dimen/spacing_small"
         android:text="@string/acknowledge_button"
-        app:layout_constraintTop_toBottomOf="@id/scrollview"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/scrollview"
         tools:text="@string/acknowledge_button" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
